Add files via upload
Creating Table using CONSTRAINTS.
CONSTRAINTS: Those are used to provide the rules for an a data.
there are 6 types of constraints.
DEFAULT- It is used for to set the default value for the columns. In a table, one or more default can be present.
CHECK-  It is used to provide the condition. In a table , one or more check  can be present.
NOT NULL- It will accept only not null values. It will accept duplicate values. It will not accept the null values. In a  table, one or more not null can be present.
UNIQUE- In a table, it will accept only unique values. It will accept null values. It will not accept duplicate values. In a tables, one or more unique can be present. 
PRIMARY KEY- It is a column, who determine the particular record uniquely. It is a combination of unique & not null. It will not accept null values and it will not accept duplicate values. In a table, only one column declare as a primary key.
FOREIGN KEY- It is a column ,in a table who refer the data of another table column. It will accept duplicate values, it will accept null values. we cannot insert any data which is not present referred table column. In a table, one or more foreign key can be present.
CREATE TABLE TABLENAME(
);